+++ src/cloud/components/molecules/CommentList.tsx
@@ -1,48 +1,31 @@
-import React, { useState, useCallback } from 'react'
+import React from 'react'
 import { Comment } from '../../interfaces/db/comments'
-import Button from '../../../shared/components/atoms/Button'
 import styled from '../../../shared/lib/styled'
-import Flexbox from '../atoms/Flexbox'
 import UserIcon from '../atoms/UserIcon'
 import { format } from 'date-fns'
 
 interface CommentThreadProps {
   comments: Comment[]
+  className: string
 }
 
-function CommentList({ comments }: CommentThreadProps) {
+// unnecessary, move up and rename file to CommentItem
+function CommentList({ comments, className }: CommentThreadProps) {
   return (
-    <ThreadContainer>
-      <div>
-        <div className='comment__list'>
-          {comments.map((comment) => (
-            <CommentItem key={comment.id} comment={comment} />
-          ))}
-        </div>
-        <CommentInput onSubmit={(cm) => console.log(cm)} />
-      </div>
-    </ThreadContainer>
+    <div className={className}>
+      {comments.map((comment) => (
+        <CommentItem key={comment.id} comment={comment} />
+      ))}
+    </div>
   )
 }
-
-const ThreadContainer = styled.div`
-  max-height: 100%;
-  overflow: auto;
-  display: flex;
-  flex-direction: column-reverse;
-  & .comment__list {
-    & > div {
-      margin-bottom: ${({ theme }) => theme.sizes.spaces.df}px;
-    }
-  }
-`
 
 interface CommentItemProps {
   comment: Comment
 }
 
 const smallUserIconStyle = { width: '24px', height: '24px', lineHeight: '20px' }
-function CommentItem({ comment }: CommentItemProps) {
+export function CommentItem({ comment }: CommentItemProps) {
   return (
     <CommentItemContainer>
       <div className='comment__meta'>
@@ -70,58 +53,6 @@ const CommentItemContainer = styled.div`
 
   & .comment__message {
     white-space: pre-wrap;
-  }
-`
-
-interface CommentInputProps {
-  onSubmit: (comment: string) => any
-}
-
-function CommentInput({ onSubmit }: CommentInputProps) {
-  const [comment, setComment] = useState('')
-
-  const onKeyDown: React.KeyboardEventHandler<HTMLTextAreaElement> = useCallback(
-    (ev) => {
-      if (ev.key === 'Enter') {
-        ev.preventDefault()
-        ev.stopPropagation()
-
-        if (ev.shiftKey) {
-          setComment((val) => `${val}\n`)
-        } else {
-          onSubmit(comment)
-        }
-      }
-    },
-    [comment, onSubmit]
-  )
-
-  const onChange: React.ChangeEventHandler<HTMLTextAreaElement> = useCallback(
-    (ev) => {
-      setComment(ev.target.value)
-    },
-    []
-  )
-
-  return (
-    <InputContainer>
-      <textarea value={comment} onChange={onChange} onKeyDown={onKeyDown} />
-      <Flexbox justifyContent='flex-end'>
-        <Button onClick={() => onSubmit(comment)}>Post</Button>
-      </Flexbox>
-    </InputContainer>
-  )
-}
-
-const InputContainer = styled.div`
-  width: 100%;
-  & textarea {
-    resize: none;
-    width: 100%;
-    border: 1px solid ${({ theme }) => theme.colors.border.main};
-    height: 60px;
-    background-color: ${({ theme }) => theme.colors.background.secondary};
-    color: white;
   }
 `
